Add Immutability class
The Immutability class assists with determining whether a specific type is immutable. Supports primitive types and certain built-in types out of the box. Allows to declare a class/struct immutable by annotating it with the [Immutable] attribute or by directly adding it to the evaluation cache of the Immutability class using the AddImmutableType<T>() method.