Add HasImmutableDerivationsOnly() method to Immutability class
This method checks whether the specified type and all derived types are immutable.
This is especially interesting when trying to find out whether a field of a certain type can be assigned immutable objects only.
Immutable objects are inherently thread-safe, so defensive copying can be avoided as these objects cannot change.

The check returns 'true', if...

a) the type is internal, i.e. it is visible within its own assembly only, and all derived types are immutable
b) the type is public, i.e. it is visible everywhere, but it does not have public, protected or protected internal constructors
   (derived types cannot be instantiated)

... otherwise the check returns 'false'.