Fix concurrency issue when getting a converter for an enumeration type
The conversion system supports enumeration types out of the box, but needs to dynamically create a specific converter for specific enumeration types. Multiple threads could try to get a converter for the same enumeration type, find out that there is no converter for that type and try to register a new converter for it. Trying to register multiple converters for the same type threw an exception. A lock around the get-then-add operation protects against this issue now.