Add NativeBitmap class and related types
The NativeBitmap class is an image class backed by a native buffer. It supports explicit disposal easing memory management. The implementation is designed close to the BitmapSource class to ease integration with WPF. Along with the NativeBitmap class related types like BitmapPalette, Color and PixelFormat have been introduced. The implementation also conforms to their WPF equivalents to make integration as easy as possible.