Fix RuntimeMetadata class
The RuntimeMetadata class now supports .NET Framework 4+, .NET Core 2/3 and .NET 5+ properly.

On .NET Framework 4+ the following assemblies are considered:
- Assemblies in the Global Assembly Cache (GAC)
- Assemblies in the application's base directory or a private bin path below the base directory
- Assemblies that have been generated dynamically

On .NET Core 2/3 and .NET 5+ the following assemblies are considered:
- Assemblies that are loaded into the default assembly load context
- Assemblies that have been generated dynamically

Other assemblies, e.g. assemblies loaded into the 'LoadFrom' context, are not considered.
Therefore assemblies that are loaded from the web are not considered.